--- src/pages/Project/components/HyperlinkList/style/index.ts
@@ -1,15 +1,0 @@
-import styled from '@emotion/styled';
-
-export const S = {
-  HyperlinkList: styled.ul({
-    display: 'flex',
-    justifyContent: 'space-between',
-  }),
-
-  HyperlinkTooltipContainer: styled.div({}),
-
-  HyperlinkImage: styled.img({
-    width: '3.2rem',
-    height: '3.2rem',
-  }),
-};
