+++ src/pages/Project/components/TooltipList/style/index.ts
@@ -1,0 +1,17 @@
+import styled from '@emotion/styled';
+
+export const S = {
+  TooltipList: styled.ul({
+    display: 'flex',
+    justifyContent: 'space-between',
+
+    maxHeight: '3.2rem',
+  }),
+
+  TooltipContainer: styled.div({}),
+
+  TooltipImage: styled.img({
+    maxWidth: '3.2rem',
+    maxHeight: '3.2rem',
+  }),
+};
